+++ gtecs/alert/data/test_notices/IceCube_13311922683750_Gold.xml
