+++ gtecs/alert/data/test_notices/LVC_S190518bb_Retraction.xml
