+++ gtecs/alert/data/test_notices/Swift_876016_BAT_GRB_Pos.xml
