+++ gtecs/alert/data/test_notices/Swift_904394_BAT_GRB_Pos.xml
