+++ gtecs/alert/tests/test_database.py
@@ -4,21 +4,21 @@
 import importlib.resources as pkg_resources
 
 from gtecs.alert import database as db
-from gtecs.alert.events import Event
+from gtecs.alert.gcn import Event
 
 
 if __name__ == '__main__':
     print('~~~~~~~~~~~~~~~')
-    test_files = sorted([f for f in pkg_resources.contents('gtecs.alert.data.test_events')
+    test_files = sorted([f for f in pkg_resources.contents('gtecs.alert.data.test_notices')
                          if f.endswith('.xml')])
-    print('Found {} test events:'.format(len(test_files)))
+    print('Found {} test notices:'.format(len(test_files)))
     for test_file in test_files:
         print(' - ', test_file)
 
     for test_file in test_files:
         print('~~~~~~~~~~~~~~~')
         print('Adding to database')
-        with pkg_resources.path('gtecs.alert.data.test_events', test_file) as f:
+        with pkg_resources.path('gtecs.alert.data.test_notices', test_file) as f:
             print(f'Loading {f}')
             event = Event.from_file(f)
         event.get_skymap()
