+++ setup.py
@@ -24,7 +24,7 @@ setup(name='gtecs-alert',
       author_email='martin.dyer@sheffield.ac.uk',
       install_requires=REQUIRES,
       packages=find_namespace_packages(include=['gtecs*']),
-      package_data={'gtecs': ['alert/data/*', 'alert/data/test_events/*']},
+      package_data={'gtecs': ['alert/data/*', 'alert/data/test_notices/*']},
       scripts=glob.glob('scripts/*'),
       zip_safe=False,
       )
